add fieldname=! for partial dereferencing
The API does partial dereferencing automatically (??)
when Sym.addr is provided: for built-in types and
small structs (<= ULONG64 size) and for pointers to
known types. For such fields it places the _value_ into
'address' field, so the offset in the struct is lost.
With fieldname=* we avoid this by placing zero to Sym.addr,
with fieldname=! we populate Sym.addr if we have it.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>